+++ src/state/entity/sc_PlayerBaseEntity-base.ts
@@ -37,31 +37,6 @@ export function setEntityState(this: ig.ENTITY.Player | sc.PartyMemberEntity, st
         this.model.spLevel = state.spLevel
     }
 
-    /* footstep sounds */
-    function getSoundFromColl(coll: ig.CollEntry, type: keyof typeof sc.ACTOR_SOUND): sc.ACTOR_SOUND_BASE {
-        var c = ig.terrain.getTerrain(coll, true, true),
-            e = sc.ACTOR_SOUND[type] || sc.ACTOR_SOUND.none
-        return (e as any)[c] ?? e[ig.TERRAIN_DEFAULT]
-    }
-    if (
-        !this.jumping &&
-        !this.animationFixed &&
-        this.stepFx.frames &&
-        !Vec2.isZero(this.coll.accelDir) &&
-        this.coll.relativeVel >= ig.ACTOR_RUN_THRESHOLD
-    ) {
-        const frame = this.animState.getFrame()
-        if (frame != this.stepFx.lastFrame) {
-            const sound = getSoundFromColl(this.coll, this.soundType)
-            if (frame == this.stepFx.frames[0]) {
-                ig.SoundHelper.playAtEntity(sound.step1!, this, null, null, 700)
-            } else if (frame == this.stepFx.frames[1]) {
-                ig.SoundHelper.playAtEntity(sound.step2!, this, null, null, 700)
-            }
-            this.stepFx.lastFrame = frame
-        }
-    } else this.stepFx.lastFrame = -1
-
     // prettier-ignore
     {
         if (state.head !== undefined) { updateStats = true; this.model.equip.head = state.head }