+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Disallow opening the quick menu on remote when player is performing action
 - Fix no door opening sound on remote
 - Fix no jumping sounds on remote
+- Fix footstep sounds playing when dashing on remote
 
 ## [0.8.3] 2026-08-01